+++ r4/src/test/java/io/github/jy95/fds/r4/translators/RateQuantityTest.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Locale;
 
 public class RateQuantityTest extends AbstractRateQuantityTest<FDSConfigR4, Dosage> {
-    
+
     @Override
     protected Dosage generateSimpleRateQuantity() {
         Dosage dosage = new Dosage();
